Attendees fix when fetching Meetup Events
- Added a getUserIfMentionable to Core Helpers which should return empty
  instead of breaking if a user can no longer be reached (e.g has left
  the server or deleted their account)
- Added a comment so I can remember how the date parsing works when it
  inevitably breaks again
- Save the events file when items are removed from the list, not sure
  why I left that out

+++ src/main/java/meetup/SeleniumDriver.java
@@ -92,7 +92,7 @@ public class SeleniumDriver {
 
 	public long sendCode(String meetupUrl, String code) {
 		lock();
-		long meetupId = new Long(extractIdFromMeetupUrl(meetupUrl));
+		long meetupId = new Long(extractIdFromMeetupUrl(meetupUrl)).longValue();
 		try {
 			webDriver.get(meetupUrl);
 			Element("//a[contains(@id, 'message-link')]").click();
